feat: Implementação completa do backend e Dockerfiles

+++ src/main/java/com/example/biblioteca/repository/AssuntoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AssuntoRepository extends JpaRepository<Assunto, Integer> {
+public interface AssuntoRepository extends JpaRepository<Assunto, Long> {
 }
